send result to server update database create and test web without logic

+++ Tests/Result/IResult.cs
@@ -2,6 +2,6 @@
 {
     internal interface IResult
     {
-        long Time { get; set; }
+        int Time { get; set; }
     }
 }
